Nested Addon Usage
Hey there,

We've been trying to use `ember-inline-svg` as a nested addon but came across an issue of `app` not being passed correctly. There is a discussion on ember-cli about it but at the moment it seems that there are only work arounds https://github.com/ember-cli/ember-cli/issues/3718

This is a small fix which sets app through the included hook.